Acted on pr feedback
