+++ src/main/java/se/sundsvall/messaging/api/HistoryResource.java
@@ -2,6 +2,7 @@ package se.sundsvall.messaging.api;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.APPLICATION_PROBLEM_JSON_VALUE;
+import static org.springframework.http.ResponseEntity.notFound;
 import static org.springframework.http.ResponseEntity.ok;
 import static se.sundsvall.messaging.Constants.BATCH_STATUS_PATH;
 import static se.sundsvall.messaging.Constants.CONVERSATION_HISTORY_PATH;
@@ -9,9 +10,6 @@ import static se.sundsvall.messaging.Constants.DELIVERY_STATUS_PATH;
 import static se.sundsvall.messaging.Constants.MESSAGE_AND_DELIVERY_PATH;
 import static se.sundsvall.messaging.Constants.MESSAGE_ATTACHMENT_PATH;
 import static se.sundsvall.messaging.Constants.MESSAGE_STATUS_PATH;
-import static se.sundsvall.messaging.Constants.STATISTICS_FOR_DEPARTMENTS_PATH;
-import static se.sundsvall.messaging.Constants.STATISTICS_FOR_SPECIFIC_DEPARTMENT_PATH;
-import static se.sundsvall.messaging.Constants.STATISTICS_PATH;
 import static se.sundsvall.messaging.Constants.USER_MESSAGES_PATH;
 import static se.sundsvall.messaging.api.model.ApiMapper.toMessageBatchResult;
 import static se.sundsvall.messaging.api.model.ApiMapper.toMessageResult;
@@ -19,7 +17,7 @@ import static se.sundsvall.messaging.api.model.ApiMapper.toMessageResult;
 import java.time.LocalDate;
 import java.util.List;
 
-import jakarta.validation.Valid;
+import jakarta.servlet.http.HttpServletResponse;
 
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
@@ -34,19 +32,12 @@ import org.zalando.problem.violations.ConstraintViolationProblem;
 import se.sundsvall.dept44.common.validators.annotation.ValidMunicipalityId;
 import se.sundsvall.dept44.common.validators.annotation.ValidUuid;
 import se.sundsvall.messaging.api.model.ApiMapper;
-import se.sundsvall.messaging.api.model.request.UserMessagesRequest;
-import se.sundsvall.messaging.api.model.response.Attachment;
 import se.sundsvall.messaging.api.model.response.DeliveryResult;
 import se.sundsvall.messaging.api.model.response.HistoryResponse;
 import se.sundsvall.messaging.api.model.response.MessageBatchResult;
 import se.sundsvall.messaging.api.model.response.MessageResult;
 import se.sundsvall.messaging.api.model.response.UserMessages;
-import se.sundsvall.messaging.api.validation.ValidNullOrNotEmpty;
-import se.sundsvall.messaging.model.DepartmentStatistics;
-import se.sundsvall.messaging.model.MessageType;
-import se.sundsvall.messaging.model.Statistics;
 import se.sundsvall.messaging.service.HistoryService;
-import se.sundsvall.messaging.service.StatisticsService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -64,16 +55,12 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 	@ApiResponse(responseCode = "400", description = "Bad Request", content = @Content(schema = @Schema(oneOf = {Problem.class, ConstraintViolationProblem.class}))),
 	@ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content(schema = @Schema(implementation = Problem.class)))
 })
-class StatusAndHistoryResource {
+class HistoryResource {
 
 	private final HistoryService historyService;
 
-	private final StatisticsService statisticsService;
-
-	StatusAndHistoryResource(final HistoryService historyService,
-		final StatisticsService statisticsService) {
+	HistoryResource(final HistoryService historyService) {
 		this.historyService = historyService;
-		this.statisticsService = statisticsService;
 	}
 
 	@Operation(summary = "Get the entire conversation history for a given party")
@@ -90,14 +77,14 @@ class StatusAndHistoryResource {
 	}
 
 	@Operation(summary = "Get the status for a message batch, its messages and their deliveries")
-
+	@ApiResponse(responseCode = "404", description = "Not Found", content = @Content(schema = @Schema(implementation = Problem.class)))
 	@GetMapping(value = BATCH_STATUS_PATH, produces = {APPLICATION_JSON_VALUE, APPLICATION_PROBLEM_JSON_VALUE})
 	ResponseEntity<MessageBatchResult> getBatchStatus(
 		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
 		@Parameter(schema = @Schema(format = "uuid")) @PathVariable @ValidUuid final String batchId) {
 
 		final var history = historyService.getHistoryByMunicipalityIdAndBatchId(municipalityId, batchId);
-		return history.isEmpty() ? ResponseEntity.notFound().build() : ok(toMessageBatchResult(history));
+		return history.isEmpty() ? notFound().build() : ok(toMessageBatchResult(history));
 	}
 
 	@Operation(summary = "Get the status for a single message and its deliveries")
@@ -108,7 +95,7 @@ class StatusAndHistoryResource {
 		@Parameter(schema = @Schema(format = "uuid")) @PathVariable @ValidUuid final String messageId) {
 
 		final var history = historyService.getHistoryByMunicipalityIdAndMessageId(municipalityId, messageId);
-		return history.isEmpty() ? ResponseEntity.notFound().build() : ok(toMessageResult(history));
+		return history.isEmpty() ? notFound().build() : ok(toMessageResult(history));
 	}
 
 	@Operation(summary = "Get the status for a single delivery")
@@ -121,7 +108,7 @@ class StatusAndHistoryResource {
 		return historyService.getHistoryByMunicipalityIdAndDeliveryId(municipalityId, deliveryId)
 			.map(ApiMapper::toDeliveryResult)
 			.map(ResponseEntity::ok)
-			.orElseGet(() -> ResponseEntity.notFound().build());
+			.orElseGet(() -> notFound().build());
 	}
 
 	@Operation(summary = "Get a message and all its deliveries")
@@ -135,49 +122,29 @@ class StatusAndHistoryResource {
 			.map(ApiMapper::toHistoryResponse)
 			.toList();
 
-		return history.isEmpty() ? ResponseEntity.notFound().build() : ok(history);
+		return history.isEmpty() ? notFound().build() : ok(history);
 	}
 
-	@Operation(summary = "Get delivery statistics")
-	@GetMapping(value = STATISTICS_PATH, produces = {APPLICATION_JSON_VALUE, APPLICATION_PROBLEM_JSON_VALUE})
-	ResponseEntity<Statistics> getStatistics(
-		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
-		@RequestParam(name = "messageType", required = false) @Parameter(description = "Message type") final MessageType messageType,
-		@RequestParam(name = "from", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Parameter(description = "From-date (inclusive). Format: yyyy-MM-dd (ISO8601)") final LocalDate from,
-		@RequestParam(name = "to", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Parameter(description = "To-date (inclusive). Format: yyyy-MM-dd (ISO8601)") final LocalDate to) {
-
-		return ok(statisticsService.getStatistics(messageType, from, to, municipalityId));
-	}
-
-	@Operation(summary = "Get letter delivery statistics by department")
-	@GetMapping(value = {STATISTICS_FOR_DEPARTMENTS_PATH, STATISTICS_FOR_SPECIFIC_DEPARTMENT_PATH}, produces = {APPLICATION_JSON_VALUE, APPLICATION_PROBLEM_JSON_VALUE})
-	ResponseEntity<List<DepartmentStatistics>> getDepartmentStatistics(
-		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
-		@PathVariable(name = "department", required = false) @Parameter(description = "Department name") @ValidNullOrNotEmpty final String department,
-		@RequestParam(name = "origin", required = false) @Parameter(description = "Origin name") final String origin,
-		@RequestParam(name = "from", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Parameter(description = "From-date (inclusive). Format: yyyy-MM-dd (ISO8601)") final LocalDate from,
-		@RequestParam(name = "to", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Parameter(description = "To-date (inclusive). Format: yyyy-MM-dd (ISO8601)") final LocalDate to) {
-
-		return ok(statisticsService.getDepartmentLetterStatistics(origin, department, from, to, municipalityId));
-	}
-
-	@Operation(summary = "Get messages sent by a user")
+	@Operation(summary = "Get historical messages sent by a user")
 	@GetMapping(value = USER_MESSAGES_PATH, produces = {APPLICATION_JSON_VALUE, APPLICATION_PROBLEM_JSON_VALUE})
-	ResponseEntity<UserMessages> getUserMessages(final @Valid UserMessagesRequest request,
-		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId) {
+	ResponseEntity<UserMessages> getUserMessages(
+		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @PathVariable @ValidMunicipalityId final String municipalityId,
+		@Parameter(name = "userId", description = "User id", example = "2281") @PathVariable(name = "userId") final String userId,
+		@Parameter(name = "page", description = "Which page to fetch", example = "1") @RequestParam(defaultValue = "1") final Integer page,
+		@Parameter(name = "page", description = "Sets the amount of entries per page", example = "1") @RequestParam(defaultValue = "15") final Integer limit) {
 
-		var result = historyService.getUserMessages(request, municipalityId);
+		var result = historyService.getUserMessages(municipalityId, userId, page, limit);
 		return ok(result);
 	}
 
-	@Operation(summary = "Get attachment by messageId and fileName")
+	@Operation(summary = "Strean attachment by messageId and fileName")
 	@GetMapping(value = MESSAGE_ATTACHMENT_PATH, produces = {APPLICATION_JSON_VALUE, APPLICATION_PROBLEM_JSON_VALUE})
-	ResponseEntity<Attachment> getAttachment(
+	void readAttachment(
 		@Parameter(name = "municipalityId", description = "Municipality id", example = "2281") @ValidMunicipalityId @PathVariable final String municipalityId,
 		@Parameter(schema = @Schema(format = "uuid")) @PathVariable @ValidUuid final String messageId,
-		@PathVariable final String fileName) {
+		@PathVariable final String fileName,
+		final HttpServletResponse response) {
 
-		var result = historyService.getAttachment(municipalityId, messageId, fileName);
-		return ok(result);
+		historyService.streamAttachment(municipalityId, messageId, fileName, response);
 	}
 }
