+++ src/main/java/se/sundsvall/messaging/service/HistoryService.java
@@ -4,10 +4,10 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
+import jakarta.servlet.http.HttpServletResponse;
+
 import org.springframework.stereotype.Service;
 
-import se.sundsvall.messaging.api.model.request.UserMessagesRequest;
-import se.sundsvall.messaging.api.model.response.Attachment;
 import se.sundsvall.messaging.api.model.response.UserMessages;
 import se.sundsvall.messaging.integration.db.DbIntegration;
 import se.sundsvall.messaging.model.History;
@@ -38,13 +38,12 @@ public class HistoryService {
 		return dbIntegration.getHistory(municipalityId, partyId, from, to);
 	}
 
-	public UserMessages getUserMessages(final UserMessagesRequest request, final String municipalityId) {
+	public UserMessages getUserMessages(final String municipalityId, final String userId, final Integer page, final Integer limit) {
 		//TODO: Implement
 		return null;
 	}
 
-	public Attachment getAttachment(final String municipalityId, final String messageId, final String fileName) {
+	public void streamAttachment(final String municipalityId, final String messageId, final String fileName, final HttpServletResponse response) {
 		//TODO: Implement
-		return null;
 	}
 }
