+++ src/test/java/se/sundsvall/messaging/TestDataFactory.java
@@ -27,9 +27,7 @@ import se.sundsvall.messaging.api.model.request.SlackRequest;
 import se.sundsvall.messaging.api.model.request.SmsBatchRequest;
 import se.sundsvall.messaging.api.model.request.SmsRequest;
 import se.sundsvall.messaging.api.model.request.SnailMailRequest;
-import se.sundsvall.messaging.api.model.request.UserMessagesRequest;
 import se.sundsvall.messaging.api.model.request.WebMessageRequest;
-import se.sundsvall.messaging.api.model.response.Attachment;
 import se.sundsvall.messaging.api.model.response.UserMessage;
 import se.sundsvall.messaging.api.model.response.UserMessages;
 import se.sundsvall.messaging.integration.emailsender.EmailDto;
@@ -425,21 +423,6 @@ public final class TestDataFactory {
 			.withCount(10)
 			.withTotalRecords(100)
 			.withTotalPages(10);
-	}
-
-
-	public static Attachment createAttachment() {
-		return Attachment.builder()
-			.withContent("someContent")
-			.withContentType("someContentType")
-			.withFileName("someFileName")
-			.build();
-	}
-
-	public static UserMessagesRequest createUserMessagesRequest() {
-		return UserMessagesRequest.builder()
-			.withUserId("someUserId")
-			.build();
 	}
 
 	public static UserMessages createUserMessages() {
