+++ src/test/java/se/sundsvall/messaging/api/HistoryResourceFailureTest.java
@@ -12,43 +12,33 @@ import static se.sundsvall.messaging.Constants.DELIVERY_STATUS_PATH;
 import static se.sundsvall.messaging.Constants.MESSAGE_AND_DELIVERY_PATH;
 import static se.sundsvall.messaging.Constants.MESSAGE_ATTACHMENT_PATH;
 import static se.sundsvall.messaging.Constants.MESSAGE_STATUS_PATH;
-import static se.sundsvall.messaging.Constants.STATISTICS_FOR_SPECIFIC_DEPARTMENT_PATH;
-import static se.sundsvall.messaging.Constants.STATISTICS_PATH;
 import static se.sundsvall.messaging.Constants.USER_MESSAGES_PATH;
-import static se.sundsvall.messaging.TestDataFactory.createUserMessagesRequest;
 
 import java.util.Map;
 import java.util.UUID;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
-import org.zalando.problem.Problem;
 import org.zalando.problem.violations.ConstraintViolationProblem;
 import org.zalando.problem.violations.Violation;
 
 import se.sundsvall.messaging.Application;
 import se.sundsvall.messaging.service.HistoryService;
-import se.sundsvall.messaging.service.StatisticsService;
 import se.sundsvall.messaging.test.annotation.UnitTest;
 
 @SpringBootTest(classes = Application.class, webEnvironment = RANDOM_PORT)
 @UnitTest
-class StatusAndHistoryResourceFailureTest {
+class HistoryResourceFailureTest {
 
 	private static final String MUNICIPALITY_ID = "2281";
 
 	@MockBean
 	private HistoryService mockHistoryService;
-
-	@MockBean
-	private StatisticsService mockStatisticsService;
 
 	@Autowired
 	private WebTestClient webTestClient;
@@ -74,7 +64,7 @@ class StatusAndHistoryResourceFailureTest {
 			.extracting(Violation::getField, Violation::getMessage)
 			.containsExactly(tuple("getConversationHistory.partyId", "not a valid UUID"));
 
-		verifyNoInteractions(mockHistoryService, mockStatisticsService);
+		verifyNoInteractions(mockHistoryService);
 	}
 
 	@Test
@@ -98,7 +88,7 @@ class StatusAndHistoryResourceFailureTest {
 			.extracting(Violation::getField, Violation::getMessage)
 			.containsExactly(tuple("getBatchStatus.batchId", "not a valid UUID"));
 
-		verifyNoInteractions(mockHistoryService, mockStatisticsService);
+		verifyNoInteractions(mockHistoryService);
 	}
 
 	@Test
@@ -122,7 +112,7 @@ class StatusAndHistoryResourceFailureTest {
 			.extracting(Violation::getField, Violation::getMessage)
 			.containsExactly(tuple("getMessageStatus.messageId", "not a valid UUID"));
 
-		verifyNoInteractions(mockHistoryService, mockStatisticsService);
+		verifyNoInteractions(mockHistoryService);
 	}
 
 	@Test
@@ -146,7 +136,7 @@ class StatusAndHistoryResourceFailureTest {
 			.extracting(Violation::getField, Violation::getMessage)
 			.containsExactly(tuple("getDeliveryStatus.deliveryId", "not a valid UUID"));
 
-		verifyNoInteractions(mockHistoryService, mockStatisticsService);
+		verifyNoInteractions(mockHistoryService);
 	}
 
 	@Test
@@ -170,67 +160,21 @@ class StatusAndHistoryResourceFailureTest {
 			.extracting(Violation::getField, Violation::getMessage)
 			.containsExactly(tuple("getMessage.messageId", "not a valid UUID"));
 
-		verifyNoInteractions(mockHistoryService, mockStatisticsService);
+		verifyNoInteractions(mockHistoryService);
 	}
 
-	@Test
-	void getStatisticsShouldFailWithInvalidMessageType() {
-		// Act
-		final var response = webTestClient.get()
-			.uri(uriBuilder -> uriBuilder
-				.path(STATISTICS_PATH)
-				.queryParam("messageType", "not-valid-type")
-				.build(Map.of("municipalityId", MUNICIPALITY_ID)))
-			.exchange()
-			.expectStatus().isBadRequest()
-			.expectHeader().contentType(APPLICATION_PROBLEM_JSON)
-			.expectBody(Problem.class)
-			.returnResult()
-			.getResponseBody();
-
-		// Assert and verify
-		assertThat(response).isNotNull();
-		assertThat(response.getDetail()).isEqualTo("""
-			Failed to convert value of type 'java.lang.String' to required type 'se.sundsvall.messaging.model.MessageType'; \
-			Failed to convert from type [java.lang.String] to type [@org.springframework.web.bind.annotation.RequestParam \
-			@io.swagger.v3.oas.annotations.Parameter se.sundsvall.messaging.model.MessageType] for value [not-valid-type]""");
-
-		verifyNoInteractions(mockHistoryService, mockStatisticsService);
-	}
-
-	@ParameterizedTest
-	@ValueSource(strings = {" ", " department", "department "})
-	void getStatisticsForDepartmentsShouldFailWithInvalidDepartment(final String department) {
-
-		// Act
-		final var response = webTestClient.get()
-			.uri(uriBuilder -> uriBuilder.path(STATISTICS_FOR_SPECIFIC_DEPARTMENT_PATH)
-				.build(Map.of("department", department, "municipalityId", MUNICIPALITY_ID)))
-			.exchange()
-			.expectStatus().isBadRequest()
-			.expectHeader().contentType(APPLICATION_PROBLEM_JSON)
-			.expectBody(ConstraintViolationProblem.class)
-			.returnResult()
-			.getResponseBody();
-
-		// Assert and verify
-		assertThat(response).isNotNull();
-		assertThat(response.getViolations())
-			.extracting(Violation::getField, Violation::getMessage)
-			.containsExactly(tuple("getDepartmentStatistics.department", "text is not null or not empty or has starting or trailing spaces"));
-
-		verifyNoInteractions(mockHistoryService, mockStatisticsService);
-	}
 
 	@Test
 	void getUserMessages_invalid_municipalityId() {
 		var municipalityId = "not-a-valid-municipalityId";
-		var userMessagesRequest = createUserMessagesRequest();
+		var userId = "userId";
+		var page = 1;
+		var limit = 15;
 
 		var response = webTestClient.get()
 			.uri(uriBuilder -> uriBuilder.path(USER_MESSAGES_PATH)
-				.queryParams(createParameterMap(userMessagesRequest.getPage(), userMessagesRequest.getLimit(), userMessagesRequest.getUserId()))
-				.build(Map.of("municipalityId", municipalityId)))
+				.queryParams(createParameterMap(page, limit))
+				.build(Map.of("municipalityId", municipalityId, "userId", userId)))
 			.exchange()
 			.expectStatus().isBadRequest()
 			.expectBody(ConstraintViolationProblem.class)
@@ -246,31 +190,7 @@ class StatusAndHistoryResourceFailureTest {
 	}
 
 	@Test
-	void getUserMessages_blank_userId() {
-		var municipalityId = "2281";
-		var userMessagesRequest = createUserMessagesRequest();
-		userMessagesRequest.setUserId("");
-
-		var response = webTestClient.get()
-			.uri(uriBuilder -> uriBuilder.path(USER_MESSAGES_PATH)
-				.queryParams(createParameterMap(userMessagesRequest.getPage(), userMessagesRequest.getLimit(), userMessagesRequest.getUserId()))
-				.build(Map.of("municipalityId", municipalityId)))
-			.exchange()
-			.expectStatus().isBadRequest()
-			.expectBody(ConstraintViolationProblem.class)
-			.returnResult()
-			.getResponseBody();
-
-		assertThat(response).isNotNull();
-		assertThat(response.getViolations())
-			.extracting(Violation::getField, Violation::getMessage)
-			.containsExactly(tuple("userId", "must not be blank"));
-
-		verifyNoInteractions(mockHistoryService);
-	}
-
-	@Test
-	void getAttachment_invalid_municipalityId() {
+	void readAttachment_invalid_municipalityId() {
 		var municipalityId = "not-a-valid-municipalityId";
 		var messageId = UUID.randomUUID().toString();
 		var fileName = "file.txt";
@@ -287,12 +207,12 @@ class StatusAndHistoryResourceFailureTest {
 		assertThat(response).isNotNull();
 		assertThat(response.getViolations())
 			.extracting(Violation::getField, Violation::getMessage)
-			.containsExactly(tuple("getAttachment.municipalityId", "not a valid municipality ID"));
+			.containsExactly(tuple("readAttachment.municipalityId", "not a valid municipality ID"));
 		verifyNoInteractions(mockHistoryService);
 	}
 
 	@Test
-	void getAttachment_invalid_messageId() {
+	void readAttachment_invalid_messageId() {
 		var municipalityId = "2281";
 		var messageId = "not-a-valid-messageId";
 		var fileName = "file.txt";
@@ -309,16 +229,15 @@ class StatusAndHistoryResourceFailureTest {
 		assertThat(response).isNotNull();
 		assertThat(response.getViolations())
 			.extracting(Violation::getField, Violation::getMessage)
-			.containsExactly(tuple("getAttachment.messageId", "not a valid UUID"));
+			.containsExactly(tuple("readAttachment.messageId", "not a valid UUID"));
 		verifyNoInteractions(mockHistoryService);
 	}
 
-	private MultiValueMap<String, String> createParameterMap(final Integer page, final Integer limit, final String userId) {
+	private MultiValueMap<String, String> createParameterMap(final Integer page, final Integer limit) {
 		MultiValueMap<String, String> parameters = new LinkedMultiValueMap<>();
 
 		ofNullable(page).ifPresent(p -> parameters.add("page", p.toString()));
 		ofNullable(limit).ifPresent(p -> parameters.add("limit", p.toString()));
-		ofNullable(userId).ifPresent(p -> parameters.add("userId", p));
 
 		return parameters;
 	}
